Update patch set 16
Patch Set 16:

(13 comments)

I gave another round of reviews (not comprehensive).

I really would like to see this patch reduced in size as there 
seems to be a lot of noise when reading it.

I'm mainly concerned about the ease of maintenance with this
work. This falls squarely in my realm of responsibility (netvirt)
and if/when things go wrong (they always do)  :)  I will probably
be someone looking to fix it.

Patch-set: 16