Update patch set 3
Patch Set 3: Code-Review+1

Here is a link to the builds that I did with this patch. 

All looks good. 

There was 1 failure in the netvirt-csit-1node-openstack-newton-nodl-v2-upstream-stateful-nitrogen job, not related to this patch.

There were 2 additional failures in the netvirt-csit-1node-openstack-newton-upstream-stateful-boron job, not related to this patch.

Patch looks good to me, ship it!

https://jenkins.opendaylight.org/sandbox/user/vpickard/builds

Patch-set: 3
Label: Code-Review=+1