Update patch set 8
Patch Set 8:

On a second thought I think test-type (functional, performance) does not introduce any extra information so you can probably neglect this for now and assume that if plot key is empty it is functional test plot, and if plot key exists it is perf plot. With that you should get plots like: openflowplugin-csit-1node-periodic-bulkomatic-perf-daily-only-nitrogen, openflowplugin-csit-1node-periodic-bulkomatic-perf-daily-only-nitrogen-rate and openflowplugin-csit-1node-periodic-bulkomatic-perf-daily-only-nitrogen-time.

Patch-set: 8