Update patch set 9
Patch Set 9: Code-Review+2

(2 comments)

I have added comments on possible improvements.
But I think the current code is nice, in the sense it is both symmetrical and readable,
so this looks good to me.
By the way, you should add Sandbox links to Gerrit so anyone could see the behavior.
https://logs.opendaylight.org/sandbox/jenkins091/controller-csit-3node-clustering-only-carbon/5/log.html.gz#s1-s15-t2-k2-k2-k1-k1-k1-k1-k2-k1-k2-k1-k2-k1-k4-k1-k1

Patch-set: 9
Label: Code-Review=+2