Fail test cases on Exception in karaf.log
There will be a global variable to control whether this check will
happen or not. a -v FAIL_ON_EXCEPTIONS:False (or True) on the
pybot command line can be used.

Also moved some keywords that are focused on parsing karaf.log
messages to KarafKeywords.robot. They did not seem to fit in
the OpenStackOperations.robot library

Change-Id: Icbc6b4669f0ad260e116b206387fabdb5d4f3e3c
Signed-off-by: Jamo Luhrsen <jluhrsen@redhat.com>