Update patch set 1
Patch Set 1: Code-Review-2

I know we need to work on a whitelist otherwise this would
introduce a lot of failures right now. I'm going to play with
this patch in the sandbox and get it right. I'm thinking a python
variables file, like we do with our data models variable file is
probably the better way to track the white list. I also need to
figure out the prettiest way (in robot) to do the check.

Patch-set: 1
Label: Code-Review=-2