Update patch set 2
Patch Set 2:

> Jamo,
 > The test are failing since some issue with the prompt in the PNF.
 > Should be trying only ping and not TCP/UDP for PNF use cases as of
 > now?
 > 
 > [1]https://logs.opendaylight.org/sandbox/jenkins091/netvirt-csit-1node-openstack-ocata-upstream-stateful-snat-conntrack-carbon/7/log.html.gz#s1-s1-s3-t21

The tests are failing because the "nc" tool is not returning the 
expected output. Are we sure it's not a legit netvirt bug?

here is a link [a] to the "nc" test above which is similar. The
command [b] is expecting to see "Test Server Data" in the output.
you can see a few steps above where it's setting up nc on the
other end to reply with that string.




[a] https://logs.opendaylight.org/sandbox/jenkins091/netvirt-csit-1node-openstack-ocata-upstream-stateful-snat-conntrack-carbon/7/log.html.gz#s1-s1-s3-t17-k2-k11
[b] sudo echo "Test Client Data" | nc -v -w 5 10.10.10.253 12345

Patch-set: 2
Label: Code-Review=0