Default BUNDLEFOLDER var to /opt/opendaylight
this currently will have zero effect on upstream ODL CSIT,
because we always pass this variable on the pybot command
line.

However, some other CSIT consumers (e.g. OPNFV) are seeing
lots of warnings [0] because they are not *always* passing
BUNDLEFOLDER in their robot argument list. This will remove
that error for them, while giving us a decent default of
/opt/opendaylight which is where the rpm package installs
to.

Change-Id: Ie32d50cd8f42b05378949439b1b1976348032e59
Signed-off-by: Jamo Luhrsen <jluhrsen@redhat.com>