Update patch set 5
Patch Set 5:

patch is good as is, but I wonder if the callers should just pass in the hostname rather than the ip so that a lookup does not have to be done? The hostnames could be cached in the beginning of the suite. This would also help with debugging since the hostnames are easier to read than the ip's.

Patch-set: 5