Update patch set 5
Patch Set 5:

> patch is good as is, but I wonder if the callers should just pass
 > in the hostname rather than the ip so that a lookup does not have
 > to be done? The hostnames could be cached in the beginning of the
 > suite. This would also help with debugging since the hostnames are
 > easier to read than the ip's.

not totally following. All we know are the compute node ips and we need
to get the hostname so we can tell nova which node to spin up the 
instance. Before we were ssh'ing to the ip address and grabbing the output
of hostname (which doesn't work in all envs).

Are you saying to start at the beginning of the suites and go figure out
the hostname from all the ips and then just pass the hostnames around?

Patch-set: 5
Reviewer: Gerrit User 2264 <2264@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0