Update patch set 5
Patch Set 5:

yes, get the ip to hostname mapping in the beginning of the test suite, add simple helpers to the libs to translate from one to the other. Then use the hostnames as much as possible so that the tests are easier to follow. My first step in troubleshooting most of these tests is to do the mapping manually.

Another possible benefit is removing the overhead time of making the api call. I know it is very small, but the durations add up over the total suite. neutron calls used to take longer but I think the openstack client is much faster.

Last thought we should also convert all the current vm instance calls to use the availability form also if not already that way. Most of the l2 and l3 suites simply let nova decide where to place the vms. To remove ambiguity we should use the function in your patch to deterministically know where the vm will be.

Patch-set: 5