Update patch set 6
Patch Set 6:

> yes, get the ip to hostname mapping in the beginning of the test
 > suite, add simple helpers to the libs to translate from one to the
 > other. Then use the hostnames as much as possible so that the tests
 > are easier to follow. My first step in troubleshooting most of
 > these tests is to do the mapping manually.
 > 
 > Another possible benefit is removing the overhead time of making
 > the api call. I know it is very small, but the durations add up
 > over the total suite. neutron calls used to take longer but I think
 > the openstack client is much faster.
 > 
 > Last thought we should also convert all the current vm instance
 > calls to use the availability form also if not already that way.
 > Most of the l2 and l3 suites simply let nova decide where to place
 > the vms. To remove ambiguity we should use the function in your
 > patch to deterministically know where the vm will be.

yeah, all good stuff. I'm like a broken record lately, but let's add
it to the list. I still can't even finish the CSIT expections patch [0]
we all want to get in, or the tempest debug patch [1]. Something like this
will definately help. And while we are talking about making things
better, we still should think about having a suite start state vs
end state check and fail if it's not the same, so we can find these
bugs where things are leftover when they shouldn't be.

[0] https://git.opendaylight.org/gerrit/#/c/60224/
[1] https://git.opendaylight.org/gerrit/#/c/62572/

Patch-set: 6