Update yang-model-validator test suite
Update this suite to reflect the changes that have been made
in the YANG Tools artifact yang-model-validator (previously known
as yang-system-test) since the test suite was first created.

The -p option previously denoted the directories with YANG files
that should be parsed as one big collection. However, the artifact
is meant to be used as a single YANG file validator similarly to
pyang. That is why the -p option is now used to specify directories
with YANG dependencies, i.e. YANG files which might be needed to
successfully validate the single YANG file specified as argument.

Therefore, adjust the suite to iterate through the set of YANG
files one by one while using the whole set of YANG models as
dependencies within the -p option.

Change-Id: I6225242bddf5dbf612a5e73153fe41c3ffa3f990
Signed-off-by: Igor Foltin <igor.foltin@pantheon.tech>