Update patch set 1
Patch Set 1:

> I'm thinking there's probably a better way to do this...
 > 
 > What's the final goal here? to provide a method to let us do some
 > debugging in the sandbox?

To debug single failures by isolating the test. This would be a simple job we can trigger when needed. We know tempest tests are a pain, but once we find a known failing test we can just test that one. Next piece of this patch is to make use of your pdb debugging.

I would rather get something in now that is useful for us to troubleshoot and we can work on adding something cooler next.

The existing tempest.robot is a template and I didn't know how to work with that since it wants to run all the tests.

Patch-set: 1
Label: Verified=0