Update patch set 2
Patch Set 2:

> > I'm thinking there's probably a better way to do this...
 > >
 > > What's the final goal here? to provide a method to let us do some
 > > debugging in the sandbox?
 > 
 > To debug single failures by isolating the test. This would be a
 > simple job we can trigger when needed. We know tempest tests are a
 > pain, but once we find a known failing test we can just test that
 > one. Next piece of this patch is to make use of your pdb debugging.
 > 
 > I would rather get something in now that is useful for us to
 > troubleshoot and we can work on adding something cooler next.
 > 
 > The existing tempest.robot is a template and I didn't know how to
 > work with that since it wants to run all the tests.

right, but in order to use this you need to file a int/test patch
that will use the tempest.single.robot job in the testplan file
and then edit which tempest test you want to run. It's not saving
much more than just creating a throwaway patch that simply trims
out all the non-needed tempest cases, like I was doing here:

https://git.opendaylight.org/gerrit/#/c/60224/7/csit/suites/openstack/tempest/tempest.robot

so the workflow is just the same right?

one idea I have is to add a jenkins param like TEMPEST_TEST_REGEX
defaulted to '' and then edit the current tempest.robot suite to
know if that var (passed on pybot cmdline) is not '' to just run
that single regex and be done.


well, thinking about it more, maybe I'm off. With this patch, if
I push a job to the sandbox, then I only have to modify two
jenkins params right?

TESTOPTIONS: -v TEMPEST_REGEX:<whatever it is>
TESTPLAN: netvirt-1node-tempest-openstack-sinlge.txt

is that how you were thinking?

Patch-set: 2