Fix the recent csit failure
There was a failure caused by the patch
(https://git.opendaylight.org/gerrit/#/c/63858/)
The min and max arguments were modified but the
instance names was left unchanged. This caused some
failures in today's run. The failure was due to a
non-exsting instance name in the list. This patch removes
the instance that is not relevent anymore.

Change-Id: I2e06c2d0d8d7e16eb36119812f3ca5d106c83fb0
Signed-off-by: gvrangan <venkatrangang@hcl.com>