Update patch set 2
Patch Set 2:

> (1 comment)
 > 
 > > (1 comment)
 > >
 > > should these pass?
 > Jamo,
 > I expect this to pass.
 > But the sandbox failed for these tests with a deprecated warning.
 > Any related file needs to be updated?
 > 
 > [1]https://jenkins.opendaylight.org/sandbox/job/netvirt-csit-1node-openstack-ocata-upstream-stateful-snat-conntrack-carbon/20/robot/netvirt-1node-openstack.txt/Tempest/Tempest/tempest.scenario.test_floatingip.FloatingIpSameNetwork/

I think we can ignore the deprecated warning. The tests seemed to fail for
valid reasons. SSHTimeout happened so I think connectivity to the instances
is broken. But a lot of the tempest tests seem to fail like that too.

something is not right with netvirt and/or openstack right now. we need
to figure this out.

Patch-set: 2
Label: Verified=0