Ignore pep8 E722
Existing code has started to fail on this error and causing
verify jobs to fail that have nothing to do with said existing
code. For now, we can ignore this to let work continue in
the int/test repo.

A jira ticket will be opened for someone to fix those related
erorrs (example below) and then remove this ignore = E722
statement.

Example:
./csit/libraries/AAAJsonUtils.py:208:13: E722 do not use bare except'
            except:
            ^
Change-Id: Iee989ea6f2bcc3149988eb2e430fc7246c2d97e2
Signed-off-by: Jamo Luhrsen <jluhrsen@redhat.com>