Update patch set 9
Patch Set 9:

> test-netvirt-current

sorry, unfortunately this does not work for int/test patches. It's
meant for java patches where it will create a new distribution based
on that java patch and then run CSIT using that distribution.

the proper way to do this is to push a job to the sandbox. Then
run that job and change the parameter PATCHREFSPEC to be the 
refspec of this patch. Currently we are on patch set 9 so the
refspec here is refs/changes/41/59941/9

In the meantime, I'm running this for you, which uses this same
refspec. we can see how it goes:

https://jenkins.opendaylight.org/releng/user/jluhrsen/my-views/view/netvirt%20csit/job/netvirt-csit-1node-openstack-ocata-gate-stateful-nitrogen/52/

Patch-set: 9