Update patch set 25
Patch Set 25: Code-Review+1

> wont be around for long, but sandbox with ps25:
 > 
 > https://jenkins.opendaylight.org/sandbox/job/netvirt-csit-1node-openstack-ocata-jamo-upstream-stateful-carbon/7/

Sandbox is ok, but there is one fallout from this patch.

We are not getting the pretty .html report files any longer. There will
need to be some work done to figure this out again. The new tool for
running tempest does not create the same output files (or even STDOUT
for that matter)

I'd prefer to catch up on this with another patch, if that's ok.

Patch-set: 25
Label: Code-Review=+1