Redirect ostestr stderr output to /dev/null
comments in the code, but for completeness here is the reason:
There are tons of deprecation error messages when we use
ostestr in our CSIT environment (openstack via devstack)
The robot log files are very large and one culprit is all
these deprecation warnings. If we redirect stderr to /dev/null
we should be able to ignore them. We will miss any other
errors, however.

Change-Id: Ic533493cf3f271e4004c0eaf9f0cb2d52e2be8c9
Signed-off-by: Jamo Luhrsen <jluhrsen@redhat.com>