Update patch set 10
Patch Set 10:

> Hi Jamo,
 > 
 > Thanks a lot for your help.
 > As you see in results of this job:
 > 
 > https://logs.opendaylight.org/releng/jenkins092/netvirt-csit-1node-openstack-ocata-gate-stateful-nitrogen/52/log.html.gz
 > 
 > 
 > "03 Basic Ipv6 Vpnservice" and "01 Basic Vpnservice" test suites
 > which are using "Collect VM IPv6 SLAAC Addresses" function have
 > passed 100%.
 > 
 > We have a failure in "ARP Learning" test suite, when it requests
 > FIB-entries list for booted vms. I don't see here, that this failed
 > step has used "Collect VM IPv6 SLAAC Addresses". All booted at this
 > step vms have IPv4 addresses.
 > 
 > So I think, this change is ready to be merged as well.
 > 
 > Regards

Valentina,

we have a problem. the log.html file you are looking at is old and not
from your run. The problem is that our robot log output.xml is getting
so big lately, that when the jobs are processing it to create a new
log.html it creates a kernel OOM error and it's killed. The log.html 
you linked to is from an old job #82. sorry for that trouble. we are
trying to find a solution to this.

BUT, the output.xml is right. What you can do is pull that file locally
and run:
  python -m robot.rebot output.xml

that will create the log.html to look at. I just did this and there
are some problems with your patch. Take a look and let me know if you
need help. Also, I am going to mark some more review comments. We need
to stop adding so many Log statements in our int/test .robot patches
when it's redundant info. That's one thing that will help the whole
kernel OOM killer problem.

again, sorry for the troubles and confusion.

JamO

Patch-set: 10