Reduce the number of polling iterations in WUKS
This has the potential to reduce some logs that end up in
the output.xml. We do use --removewuks when we generate the
html reports, but that is post processing. All the iterations
are still being logged in output.xml. Hopefully this will
help reduce the final log size of output.xml

Change-Id: If84c43ac659750e53efe2b2c18c6e81bab801ad7
Signed-off-by: Jamo Luhrsen <jluhrsen@redhat.com>