Reduce logging in collect ips
Created a new VM Collect IP keyword that reduces the
amount of logging and works faster.

Problems in old keyword:
- logged the whole console multiple times, per vm check and per
operation
- ran Verify Active extra times
- different suites replicated using code

This patch:
- moves all the common wrapping code to a single keyword
- reduces the extra Active calls
- reduces the console to a single time per vm check
- removed console log on error since the log is always captured
during the attempt to get the addresses already

This looks to have reduced the time in half for retrieving the
addresses as well as reducing the multiple loggin of long
console logs.

Only converted the connectivity suite and left the old keyword
in place. The next patches will convert the other suites.

Change-Id: I042c137165eae95946fa5197fbb52b1b85312a00
Signed-off-by: Sam Hague <shague@redhat.com>