Workarounds for containers
Adding NODE_KARAF_COUNT_COMMAND as all the other operations on
nodes can be overwritten on command line and this is useful when
running from a container.

Also changing the start function to check for "docker" as we don't
want to add those options when using docker.

This is just a temporary fix that makes it usable, but it could be
limiting and require more work later

Change-Id: I5e90cefcb52dec9b9666813824a6a42721db173d
Signed-off-by: Tomas Jamrisko <tjamrisk@redhat.com>