Update patch set 1
Patch Set 1: Code-Review-2

> merge this?

no. this needs to be refactored to use the new format. Also, I
think this is a higher level tempest bucket and now that we debug
on teardown() we have to run the cases one at a time.

I'll get to this eventually.

Patch-set: 1
Reviewer: Gerrit User 2264 <2264@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=-2