Update patch set 16
Patch Set 16:

> > > do we want this?
 > > >
 > > > new sb run here:
 > > > https://jenkins.opendaylight.org/sandbox/job/netvirt-csit-1node-openstack-ocata-jamo-upstream-stateful-carbon/1/
 > >
 > > Is every exception in the logs whitelisted? I can't find a place
 > > where the new test would fail.
 > 
 > Oh, the default is False so nothing happens.

I forgot about that, but that's a good test then. We know that the flag is
working. Here:

https://jenkins.opendaylight.org/sandbox/job/netvirt-csit-1node-openstack-ocata-jamo-upstream-stateful-carbon/3/

that is run with pybot having -v FAIL_ON_EXCEPTIONS:True

btw, if we are going this route we may want to take this one step further and introduce
a JJB Param for it. Then for each job/branch/etc we can toggle this check more
easily. Maybe we don't want to worry about it for Carbon, as an example.

Patch-set: 16