Update patch set 15
Patch Set 15: Code-Review-1

Jamo, this patch will be useful for the new periodic jobs in carbon and nitrogen to enable genius auto tunnels, and disable netvirt dynamic tunnels.

For oxygen, the default will be changed to genius auto tunnels, so we won't need this patch in there and later streams.

Patch-set: 15
Label: Code-Review=-1