Update patch set 1
Patch Set 1:

Yes, there is: http://docs.opendaylight.org/en/stable-boron/submodules/integration/test/docs/system-test-guide.html, the intention of starting with the basic test is for new projects to get familiar with the ODL test infra, clearly this does not count as a valid system test and projects have to work in their own test later. If this is not the way we want projects to get started we should change the system test guide.

Patch-set: 1
Label: Code-Review=0