Update patch set 15
Patch Set 15: Code-Review+2

we are going to lose visibility on https://jira.opendaylight.org/browse/NETVIRT-1019 with this patch.

I will update NETVIRT-1019 to indicate that we should not RESOLVE
it until that specific case is added to the tempest suite. It really
just just be as simple as adding a final test case that deletes
the external-net network. That will fail if there is still a port
leak and it would just be a single test failure. This patch will
then make the full openstack suite cleanup run which really will
delete that network anyway.

Patch-set: 15
Label: Code-Review=+2