Update patch set 3
Patch Set 3:

> > Jamo, do we want to merge this?
 > 
 > let's wait to see what the sandbox looks like. I had trouble
 > with it yesterday, but they are running now:
 > 
 > https://jenkins.opendaylight.org/sandbox/job/netvirt-csit-1node-openstack-pike-jamo-upstream-stateful-carbon/1/
 > https://jenkins.opendaylight.org/sandbox/job/netvirt-csit-1node-openstack-ocata-jamo-upstream-stateful-carbon/1/

the links above will be valid in the new and old sandbox. I saw
that the jobs in the new sandbox were having a lot of failures
and I'm not sure if there is some env trouble there or not yet. So,
I started them in the new sandbox as well.

Patch-set: 3
Label: Verified=0