Update patch set 13
Patch Set 13: Code-Review-1

Jamo,
Thanks for the comments. Originally, wasn't intending to merge this patch, just for debugging. But, some of this may be useful going forward, maybe we could come up with some flag to "Enable Extra Debug" or something, and just flip it on and off. And cleanup things you mentioned in your comments also as part of that, if we decide it would be helpful to have some of these going forward. Thanks Jamo!

Patch-set: 13
Label: Code-Review=-1