Update patch set 7
Patch Set 7:

> why are we doing this? just curious.

We don't use the stats collection. My understanding is this was just a way to always have the stats in ds for some applications rather than on demand. The real issue is this causes perf issues since it is every 3s per switch, all being written to ds.

Patch-set: 7
Label: Verified=0