Update patch set 7
Patch Set 7: Code-Review-1

which patch is the more recent - this one or the other? Please abandon one. Make sure to look at the comments on the other patch.

also is there a reason there are two? it is best to keep the Change-Id the same in patches if they are really the same one, otherwise we lose the history of the patch. Make sure to git commit --amend to keep the original patch.

Patch-set: 7
Reviewer: Gerrit User 773 <773@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=-1