Redirect openstack cli stderr > /dev/null
Something is slightly broken with our current setup of using
latest openstackclient 13.0.0 with one version back of
novaclient 9.0.0 such that when doing calls like "server
list" we are gettins some stderr output that looks like
some of the debug output when using -v or --debug.

this will supress it for now. Probably a bug should be
raised with openstackclient that it's not handling the
older novaclient properly. Not sure if they are supposed
to work nicely or not.

Change-Id: I0036976be7a907287452aa3b3a91228b820b3c1b
Signed-off-by: Jamo Luhrsen <jluhrsen@redhat.com>