Pop VM from list after deleting it
now that we are checking nova commands for proper return
codes, we cannot just happily delete things that are already
deleted. So, for the test case where we delete a single
instance to make sure ping traffic fails, we also need to
remove that instance from the list and let further cleanup
skip it.

Change-Id: Id0578e2c187e4b6703f76a11e369ffb92f6e17af
Signed-off-by: Jamo Luhrsen <jluhrsen@redhat.com>