Update patch set 6
Patch Set 6: Code-Review+2

external_network wasn't supposed to have the tcpdump - it was added for debugging the other problems. So it was a mistake that it was in this patch. I have now moved that out to it's own patch.

I think we should just go ahead and add tcpdump to all the suites. Looks like the files are fairly small. That way we don't have to keep adding debugging patches and pushing jobs.

Patch-set: 6
Label: Code-Review=+2