Update patch set 59
Patch Set 59: Code-Review+2

(1 comment)

putting this in the new tdd-wip csit job for now:

https://git.opendaylight.org/gerrit/67738

let's let it run and try to get momentum for people to fix the
failures. I believe we are attributing the failures to these
two jira issues:

https://jira.opendaylight.org/browse/NETVIRT-1060
https://jira.opendaylight.org/browse/NETVIRT-1061

big thanks for doing this work, Valentina.

Patch-set: 59
Label: Code-Review=+2