Update patch set 67
Patch Set 67: Code-Review-1

(12 comments)

Can we add more details in the commit message on how this works? Biggest confusion I have is this looks to be doing some of the work that the builder side was doing, so not sure what the work split is now.

Patch-set: 67
Reviewer: Gerrit User 773 <773@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=-1