Update patch set 8
Patch Set 8:

so this is working as expected. in pike the new test case fails and
we need this to get fixed. it actually gets in the way of other
testing. And it passes in ocata.

I vote to merge the test and we fight to get the sloppy bug fixed.

copy-logs: tempest-leftovers-netvirt-csit-1node-openstack-pike-upstream-stateful-nitrogen/3

Patch-set: 8