WIP: Fix L2GW CSIT suite init
L2GW CSIT only has 1 compute node, so recent changes to
openstack utilities caused l2gw csit to bomb terribly.

Not sure if this is the correct fix, will run in
sandbox when jenkins is back online to see.

This is similiar to patch https://git.opendaylight.org/gerrit/#/c/68012/,
will distill to 1 patch when we figure this out.

Change-Id: I217f97f7c66bc0089d2e932b285bc08b88e60f0f
Signed-off-by: Victor Pickard <vpickard@redhat.com>