Update patch set 4
Patch Set 4:

(1 comment)

> Patch Set 4:
> 
> This patch doesn't solve the issues we're facing in HA setup... to solve the issue we'd have to add two more variables for jolokia credentials. We'd be just seeing 401s on different places

right, but the idea of the patch is still useful. we want to be able to
override ${AUTH} in this file if needed. so we need to source Variables.robot
to make sure it's available.

Patch-set: 4
Reviewer: Gerrit User 2264 <2264@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0