Update patch set 3
Patch Set 3:

> (2 comments)

I asked Faseela to go for OVS/OpenVSwitch as it relates to switch which is/can-be common to all projects. I first thought of OVSDB, but that is specific to OVSDB project, OpenStack is rather misleading. IMO would be good to have OVS [for switch] to have all methods for OVS and have all modules, including OpenSTack, OVSDB etc. reuse from it.

Patch-set: 3