Update patch set 3
Patch Set 3:

Openstack Operations.robot has more variables which we don't use in genius, like OS_SYSTEM_IP etc. This will add up to the parameters which we have to pass when we run this locally. And having these keywords in openstack related resources did not make much sense. So thought of adding a new OVS.robot and migrating everyone else to that

Patch-set: 3