Update patch set 3
Patch Set 3: Code-Review-1

looks like you are moving files to a new folder with new
names? but also some refactoring in the files as well?

could you do this is a couple of steps so it's easier to 
review the refactoring part? this patch just makes it look
like they are all new files (the ones that got renamed)

Patch-set: 3
Label: Code-Review=-1