Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> @Jamo, @Sam: Idea here is to move these keywords to OVS.robot and have OpenStack, Genius and even OVSDB.robot reuse from that one. OVS.robot to host all switch related methods that others can need. First step is to add OVS.robot with this patch, next could be to refator openstack.robot to use OVS.robot for OVS stuff. Or are you not okay with idea of refactoring at all?

no refactoring to make things more clear and reusable is always awesome, imho.

my fear is this work is going to duplicate stuff (which is opposite of
refactoring) with the potential goal to make it better down the road.

Can we just go all the way with this patch? so move the right stuff to
the new OVS and remove the duplicate stuff, which will require an update
to all the places we are calling it from OpenStackOperations.

I know it's a little more work, but I worth it to get it right.

Sam?

Patch-set: 3