Add Variables.robot to get visibility to ${AUTH}
in the case when a suite uses this cluster management (CM)
library, but has not imported the global Variables.robot
file, certain variables are not available (e.g. ${AUTH})
which is used in this CM lib. This can cause problems
like unauthorized access when sending REST calls to the
controller.

Change-Id: Ib0bbd1577ea32a5978904e6ef483146a502398b0
Signed-off-by: Jamo Luhrsen <jluhrsen@redhat.com>