Update patch set 34
Patch Set 34:

> (2 comments)
Let me move everything to OVSDB.lib, but we should clean this stuff up over time.

These are specific to the ovsdb suites so they don't readily fit in the OVSDB.lib, If you create an ovsdb dir then that is what this patch is doing by keeping the ovsdb suite-specific keywords near the suite itself rather than dumping them in some. A big problem with the libs today is that they are dumping grounds since they are in a lib dir. Makes sense for common functions used by a number of suites but over time it starts to pollute things. Variables are even worse. Would really like to see a lot of stuff move out of these dumping grounds and put closer to the suites themselves. If something goes beyond the scope of a quite then you move it up to a higher level lib.

Patch-set: 34
Label: Verified=0