Update patch set 35
Patch Set 35:

> Patch Set 34:
> 
> > (2 comments)
> Let me move everything to OVSDB.lib, but we should clean this stuff up over time.
> 
> These are specific to the ovsdb suites so they don't readily fit in the OVSDB.lib, If you create an ovsdb dir then that is what this patch is doing by keeping the ovsdb suite-specific keywords near the suite itself rather than dumping them in some. A big problem with the libs today is that they are dumping grounds since they are in a lib dir. Makes sense for common functions used by a number of suites but over time it starts to pollute things. Variables are even worse. Would really like to see a lot of stuff move out of these dumping grounds and put closer to the suites themselves. If something goes beyond the scope of a quite then you move it up to a higher level lib.

I get what you are saying and I'll defer to you for what best practices
should be. I'm mostly commenting on how we have developed things until
now. I also hate seeing code come in that is a literal copy/paste of
something that exists somewhere else, so I default to telling people to
use what we have in csit/libraries/* first.

If it's only used once in a suite, then we suggest to write those things
in the suite itself as a local keyword. But, you bring up the point about
how to handle things that are strictly common to a few suites in a single
folder. should we start suggesting the model of csit/suites/ovsdb/lib/ or
something like that?

as for things like csit/libraries/OVSDB.robot, I know there are keywords
in there used across several projects, so that still seems like a good
idea to have. for example, the keyword to get the ovsdb UUID, etc.

Patch-set: 35
Reviewer: Gerrit User 2264 <2264@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0