Update patch set 5
Patch Set 5:

> Patch Set 5:
> 
> Jamo, this ready? How big is the trace dump or what does it look like?

not yet. I think I want to add the check (it's fast) to every test teardown
too. that will help us narrow down the leaks.

the problem I see is that unless we make this stuff failing, we will
have to remember to check. like the exception failing stuff. and if
we aren't diligent with checking gate jobs, new ones will creep in.

Patch-set: 5
Label: Verified=0