Update patch set 4
Patch Set 4:

this ran in the sandbox, but because of a lot of failures the job took longer than 6 hours to finish, so jenkins pulled the plug and we didn't get robot
logs. we do have the console logs so you can look at the output of pybot:

https://logs.opendaylight.org/releng/vex-yul-odl-jenkins-1/builder-copy-sandbox-logs/20/jamo-netvirt-csit-3node-openstack-ocata-upstream-stateful-oxygen/3/console.log.gz

there were 8 failures in this suite. is that expected?

Patch-set: 4