Update patch set 4
Patch Set 4: Code-Review-1

(5 comments)

We really should find a way to not have to duplicate 90% of a file in another suite. A starter might be to just move everything to some keywords that are shred between the suites. That at least keeps the code in one place, though you still need to dup the calls of the keywords in the tests.

Patch-set: 4
Reviewer: Gerrit User 773 <773@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=-1