Update patch set 76
Patch Set 76:

actually, can you move the libraries and suites under deployment as well?

I just want to fully separate this work until it becomes the defacto method
of deployment (which is still a long way off, if that happens) so that it
doesn't confuse the stuff we have standardized on to this point.

you'll still be able to use/leverage all of it in any new jobs coming
in.

Patch-set: 76
Label: Verified=0